Adds react-router-redux
