+++ src/containers/article.js
@@ -3,21 +3,9 @@ import ArticleFullPage from '../components/article/';
 
 import * as ArticleActions from '../actions/article';
 
-const Actions = {
-  ...ArticleActions
-};
-
 function mapStateToProps (state) {
-  const {
-    article: {
-      articlePage,
-      articleContent
-    }
-  } = state;
-  return {
-    articlePage,
-    articleContent
-  };
+  const { article: { articleContent } } = state;
+  return { articleContent };
 }
 
-export default connect(mapStateToProps, Actions)(ArticleFullPage);
+export default connect(mapStateToProps, ArticleActions)(ArticleFullPage);
