+++ src/containers/hotel.js
@@ -3,19 +3,9 @@ import HotelPage from '../components/hotel/';
 
 import * as HotelActions from '../actions/hotel';
 
-const Actions = {
-  ...HotelActions
-};
-
 function mapStateToProps (state) {
-  const {
-    hotel: {
-      hotelInView
-    }
-  } = state;
-  return {
-    packageOffer: hotelInView
-  };
+  const { hotel: { hotelInView } } = state;
+  return { packageOffer: hotelInView };
 }
 
-export default connect(mapStateToProps, Actions)(HotelPage);
+export default connect(mapStateToProps, HotelActions)(HotelPage);
