+++ src/reducers/hotel.js
@@ -1,9 +1,8 @@
 'use strict';
 
-import { VIEW_HOTEL, VIEW_SEARCH, SET_HOTEL_PAGE } from '../constants/actionTypes';
+import { SET_HOTEL_PAGE } from '../constants/actionTypes';
 
 const initialState = {
-  hotelPage: false,
   hotelInView: {
     description: '',
     hotel: {
@@ -93,16 +92,6 @@ const initialState = {
 
 export default function hotel (state = initialState, action) {
   switch (action.type) {
-    case VIEW_HOTEL:
-      return {
-        ...state,
-        hotelPage: true
-      };
-    case VIEW_SEARCH:
-      return {
-        ...state,
-        hotelPage: false
-      };
     case SET_HOTEL_PAGE:
       return {
         ...state,
